requirements.txt
groq==0.11.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
rich==13.7.0